Replaces SearchScreen component with AlgoSearchScreen
