+++ src/routes/search.js
@@ -1,5 +1,5 @@
 import { StackNavigator } from 'react-navigation';
-import SearchScreen from 'kitsu/screens/Search/AlgoSearchScreen';
+import SearchScreen from 'kitsu/screens/Search/SearchScreen';
 import SearchCategory from 'kitsu/screens/Search/SearchCategory';
 import SearchResults from 'kitsu/screens/Search/SearchResults';
 import navigationOptions from './navigationOptions';
