+++ src/screens/Search/Lists/ResultsList.js
@@ -1,107 +1,12 @@
 import React from 'react';
-import PropTypes from 'prop-types';
-import { View, FlatList, Text, TouchableOpacity } from 'react-native';
+import { View, FlatList, Text, TouchableOpacity, StyleSheet } from 'react-native';
 import LinearGradient from 'react-native-linear-gradient';
 import ProgressiveImage from 'kitsu/components/ProgressiveImage';
 
-const ResultsList = ({
-  dataArray,
-  loadMore,
-  refreshing = false,
-  onPress,
-  refresh,
-  scrollEnabled = true,
-  numColumns = 4,
-  imageSize = { h: 125, w: 91 },
-}) => (
-  <View style={{ backgroundColor: '#FAFAFA' }}>
-    <FlatList
-      removeClippedSubviews={false}
-      data={dataArray}
-      onEndReached={() => loadMore()}
-      onEndReachedThreshold={0.5}
-      getItemLayout={(data, index) => ({
-        length: imageSize.h,
-        offset: imageSize.h * index,
-        index,
-      })}
-      initialNumToRender={10}
-      numColumns={numColumns}
-      scrollEnabled={scrollEnabled}
-      refreshing={refreshing}
-      onRefresh={() => refresh()}
-      contentContainerStyle={styles.list}
-      renderItem={e => renderItem(e, imageSize, onPress)}
-    />
-  </View>
-);
+const IMAGE_SIZE = { height: 125, width: 91 };
 
-const renderItem = ({ item }, imageSize, onPress) => {
-  let title = null;
-  if (item.titles) {
-    title = item.titles.en || item.titles.en_jp;
-  }
-  const { h, w } = imageSize;
-  const m = imageSize.m || 1;
-  return (
-    <TouchableOpacity onPress={() => onPress(item)}>
-      <View
-        style={{
-          height: h - m * 2,
-          width: w - m * 2,
-          margin: m,
-        }}
-        onPress={() => onPress(item)}
-      >
-        <ProgressiveImage
-          onPress={() => onPress(item)}
-          source={{ uri: item.image }}
-          style={{
-            height: h - m * 2,
-            width: w - m * 2,
-          }}
-        />
-        {title &&
-          <LinearGradient colors={['transparent', 'black']} style={styles.linearGradient}>
-            <Text
-              style={{
-                color: 'white',
-                backgroundColor: 'transparent',
-                fontSize: 12,
-                fontFamily: 'OpenSans',
-                fontWeight: '600',
-                padding: 3,
-              }}
-              numberOfLines={2}
-            >
-              {title}
-            </Text>
-          </LinearGradient>}
-      </View>
-    </TouchableOpacity>
-  );
-};
-
-ResultsList.propTypes = {
-  dataArray: PropTypes.array.isRequired,
-  onPress: PropTypes.func,
-  loadMore: PropTypes.func,
-  refresh: PropTypes.func,
-  refreshing: PropTypes.bool,
-  scrollEnabled: PropTypes.bool,
-  imageSize: PropTypes.object,
-};
-
-ResultsList.defaultProps = {
-  loadMore: () => {},
-  refresh: () => {},
-  onPress: () => {},
-  imageSize: {},
-  refreshing: false,
-  scrollEnabled: true,
-};
-
-const styles = {
+const styles = StyleSheet.create({
+  container: { backgroundColor: '#FAFAFA' },
   list: {
     justifyContent: 'center',
     flexDirection: 'row',
@@ -116,6 +21,77 @@ const styles = {
     height: '40%',
     justifyContent: 'flex-end',
   },
+  titleText: {
+    color: 'white',
+    backgroundColor: 'transparent',
+    fontSize: 12,
+    fontFamily: 'OpenSans',
+    fontWeight: '600',
+    padding: 3,
+  },
+});
+
+const ResultsList = ({ hits, hasMore, refine, onPress }) => {
+  const onEndReached = () => {
+    if (hasMore) {
+      refine();
+    }
+  };
+  return (
+    <FlatList
+      removeClippedSubviews={false}
+      data={hits}
+      onEndReached={onEndReached}
+      onEndReachedThreshold={0.5}
+      getItemLayout={(data, index) => ({
+        length: IMAGE_SIZE.height,
+        offset: IMAGE_SIZE.height * index,
+        index,
+      })}
+      initialNumToRender={10}
+      numColumns={4}
+      scrollEnabled
+      contentContainerStyle={styles.list}
+      renderItem={e => renderItem(e, onPress)}
+      style={styles.container}
+    />
+  );
+};
+
+const renderItem = ({ item }, onPress) => {
+  let title = null;
+  if (item.titles) {
+    title = item.titles.en || item.titles.en_jp;
+  }
+  const { height, width } = IMAGE_SIZE;
+  const m = IMAGE_SIZE.m || 1;
+  return (
+    <TouchableOpacity onPress={() => onPress(item)}>
+      <View
+        style={{
+          height: height - m * 2,
+          width: width - m * 2,
+          margin: m,
+        }}
+        onPress={() => onPress(item)}
+      >
+        <ProgressiveImage
+          onPress={() => onPress(item)}
+          source={{ uri: (item.posterImage || {}).small }}
+          style={{
+            height: height - m * 2,
+            width: width - m * 2,
+          }}
+        />
+        {title &&
+          <LinearGradient colors={['transparent', 'black']} style={styles.linearGradient}>
+            <Text style={styles.titleText} numberOfLines={2}>
+              {title}
+            </Text>
+          </LinearGradient>}
+      </View>
+    </TouchableOpacity>
+  );
 };
 
 export default ResultsList;
