+++ src/screens/Search/Lists/index.js
@@ -1,6 +1,5 @@
 import ResultsList from './ResultsList';
-import AlgoResultsList from './AlgoResultsList';
 import TopsList from './TopsList';
 import UsersList from './UsersList';
 
-export { ResultsList, UsersList, TopsList, AlgoResultsList };
+export { ResultsList, UsersList, TopsList };
